+++ src/ts/dom/dom.ts
@@ -63,7 +63,7 @@ export namespace DomElement {
         return result;
     }
 
-    export function showElementAtMousePosition( e: MouseEvent, element: HTMLElement ) : void {
+    export function showElementAtMousePosition( e: MouseEvent, element: HTMLElement, offset: number ) : void {
         if ( element.style.display !== "block" ) {
             let left: number = e.pageX;
             let top: number = e.pageY;
@@ -72,15 +72,17 @@ export namespace DomElement {
             element.style.display = "block";
     
             if ( left + element.offsetWidth > window.innerWidth ) {
-                left -= element.offsetWidth;
+                left -= ( element.offsetWidth + offset );
             } else {
                 left++;
+                left += offset;
             }
     
             if ( top + element.offsetHeight > window.innerHeight ) {
-                top -= element.offsetHeight;
+                top -= ( element.offsetHeight + offset );
             } else {
                 top++;
+                top += offset;
             }
     
             if ( left < scrollPosition.left ) {