+++ src/ts/options/options.ts
@@ -29,6 +29,7 @@ export namespace Start {
             options.showNodeNameInTitle = Default.getBoolean( options.showNodeNameInTitle, false );
             options.ignoreValues = Default.getStringOrArray( options.ignoreValues, [] );
             options.showLockButtonInTitle = Default.getBoolean( options.showLockButtonInTitle, true );
+            options.dialogOffset = Default.getNumber( options.dialogOffset, 0 );
 
             return options;
         }